refactor: State abstract -> interface 로 변경
