+++ src/main/java/bowling/domain/state/finish/Miss.java
@@ -3,8 +3,9 @@ package bowling.domain.state.finish;
 import bowling.domain.score.Score;
 import bowling.domain.state.State;
 import bowling.domain.Pins;
+import bowling.exception.ImpossiblePitchException;
 
-public class Miss extends Finish {
+public class Miss implements State {
 
     private static final String GUTTER = "-";
     private static final String VERTICAL_BAR = "|";
@@ -47,6 +48,16 @@ public class Miss extends Finish {
     }
     private String convertIfGutter(Pins pins) {
         return pins.isGutter() ? GUTTER : pins.toString();
+    }
+
+    @Override
+    public boolean isFrameEnd() {
+        return true;
+    }
+
+    @Override
+    public State pitch(Pins pins) {
+        throw new ImpossiblePitchException(State.class.getName());
     }
 
 }
